pyproject.toml
[tool.poetry]
name = "ig-feed-generator"
version = "0.1.0"
description = ""
authors = ["mukund1712 <143148571+mukund1712@users.noreply.github.com>", "abodh788 abodh.gajurel100@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
pre-commit = "^3.8.0"
openai = "^1.43.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.8.0"
ruff = "^0.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
